refactor hidden duplicate interactions into own place now it exists :)
