+++ app/src/main/java/eu/kanade/domain/DomainModule.kt
@@ -66,6 +66,9 @@ import tachiyomi.domain.chapter.interactor.SetMangaDefaultChapterFlags
 import tachiyomi.domain.chapter.interactor.ShouldUpdateDbChapter
 import tachiyomi.domain.chapter.interactor.UpdateChapter
 import tachiyomi.domain.chapter.repository.ChapterRepository
+import tachiyomi.domain.hiddenDuplicates.interactor.AddHiddenDuplicate
+import tachiyomi.domain.hiddenDuplicates.interactor.GetAllHiddenDuplicates
+import tachiyomi.domain.hiddenDuplicates.interactor.RemoveHiddenDuplicate
 import tachiyomi.domain.hiddenDuplicates.repository.HiddenDuplicateRepository
 import tachiyomi.domain.history.interactor.GetHistory
 import tachiyomi.domain.history.interactor.GetNextChapters
@@ -73,7 +76,6 @@ import tachiyomi.domain.history.interactor.GetTotalReadDuration
 import tachiyomi.domain.history.interactor.RemoveHistory
 import tachiyomi.domain.history.interactor.UpsertHistory
 import tachiyomi.domain.history.repository.HistoryRepository
-import tachiyomi.domain.manga.interactor.AddHiddenDuplicate
 import tachiyomi.domain.manga.interactor.FetchInterval
 import tachiyomi.domain.manga.interactor.GetDuplicateLibraryManga
 import tachiyomi.domain.manga.interactor.GetFavorites
@@ -83,7 +85,6 @@ import tachiyomi.domain.manga.interactor.GetManga
 import tachiyomi.domain.manga.interactor.GetMangaByUrlAndSourceId
 import tachiyomi.domain.manga.interactor.GetMangaWithChapters
 import tachiyomi.domain.manga.interactor.NetworkToLocalManga
-import tachiyomi.domain.manga.interactor.RemoveHiddenDuplicates
 import tachiyomi.domain.manga.interactor.ResetViewerFlags
 import tachiyomi.domain.manga.interactor.SetMangaChapterFlags
 import tachiyomi.domain.manga.interactor.UpdateMangaNotes
@@ -123,8 +124,7 @@ class DomainModule : InjektModule {
 
         addSingletonFactory<MangaRepository> { MangaRepositoryImpl(get()) }
         addFactory { GetDuplicateLibraryManga(get()) }
-        addFactory { AddHiddenDuplicate(get()) }
-        addFactory { RemoveHiddenDuplicates(get()) }
+        addFactory { GetHiddenDuplicates(get()) }
         addFactory { GetFavorites(get()) }
         addFactory { GetLibraryManga(get()) }
         addFactory { GetMangaWithChapters(get(), get()) }
@@ -150,7 +150,9 @@ class DomainModule : InjektModule {
         }
 
         addSingletonFactory<HiddenDuplicateRepository> { HiddenDuplicateRepositoryImpl(get()) }
-        addFactory { GetHiddenDuplicates(get(), get()) }
+        addFactory { AddHiddenDuplicate(get()) }
+        addFactory { RemoveHiddenDuplicate(get()) }
+        addFactory { GetAllHiddenDuplicates(get()) }
 
         addSingletonFactory<ReleaseService> { ReleaseServiceImpl(get(), get()) }
         addFactory { GetApplicationRelease(get(), get()) }
