+++ app/src/main/java/eu/kanade/tachiyomi/data/backup/create/creators/HiddenDuplicatesBackupCreator.kt
@@ -2,16 +2,16 @@ package eu.kanade.tachiyomi.data.backup.create.creators
 
 import eu.kanade.tachiyomi.data.backup.models.BackupHiddenDuplicate
 import eu.kanade.tachiyomi.data.backup.models.backupHiddenDuplicateMapper
-import tachiyomi.domain.manga.interactor.GetHiddenDuplicates
+import tachiyomi.domain.hiddenDuplicates.interactor.GetAllHiddenDuplicates
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 
 class HiddenDuplicatesBackupCreator(
-    private val getHiddenDuplicates: GetHiddenDuplicates = Injekt.get(),
+    private val getAllHiddenDuplicates: GetAllHiddenDuplicates = Injekt.get(),
 ) {
 
     suspend operator fun invoke(): List<BackupHiddenDuplicate> {
-        return getHiddenDuplicates.await()
+        return getAllHiddenDuplicates.await()
             .map(backupHiddenDuplicateMapper)
     }
 }
