+++ domain/src/main/java/tachiyomi/domain/hiddenDuplicates/interactor/GetAllHiddenDuplicates.kt
@@ -1,0 +1,13 @@
+package tachiyomi.domain.hiddenDuplicates.interactor
+
+import tachiyomi.domain.hiddenDuplicates.repository.HiddenDuplicateRepository
+import tachiyomi.domain.manga.model.HiddenDuplicate
+
+class GetAllHiddenDuplicates(
+    private val hiddenDuplicateRepository: HiddenDuplicateRepository,
+) {
+
+    suspend fun await(): List<HiddenDuplicate> {
+        return hiddenDuplicateRepository.getAll()
+    }
+}
